# modules/fundamental/src/purl/endpoints/mod.rs
use crate::{
    Error,
    endpoints::Deprecation,
    purl::{
        model::{
            RecommendReportRequest, RecommendReportResponse, RecommendRequest, RecommendResponse,
            details::purl::PurlDetails, summary::purl::PurlSummary,
        },
        service::PurlService,
    },
};
use actix_web::{HttpResponse, Responder, get, post, web};
use regex::Regex;
use sea_orm::prelude::Uuid;
use std::str::FromStr;
use trustify_auth::{ReadAdvisory, ReadSbom, authorizer::Require};
use trustify_common::{
    db::{self, pagination_cache::PaginationCache, query::Query},
    id::IdError,
    model::{Paginated, PaginatedResults},
    purl::Purl,
};

mod base;

#[cfg(test)]
mod test;

pub fn configure(
    config: &mut utoipa_actix_web::service_config::ServiceConfig,
    db: db::ReadOnly,
    cache: PaginationCache,
    recommend_patterns: Vec<Regex>,
    report_package_limit: u64,
) {
    let purl_service = PurlService::new(cache)
        .with_recommend_patterns(recommend_patterns)
        .with_report_package_limit(report_package_limit);

    config
        .app_data(web::Data::new(db))
        .app_data(web::Data::new(purl_service))
        .service(base::get_base_purl)
        .service(base::all_base_purls)
        .service(v2::recommend) // Must be before `get` to avoid {key} matching "recommend"
        .service(v3::recommend) // Must be before `get` to avoid {key} matching "recommend"
        .service(v3::recommend_report)
        .service(all)
        .service(get);
}

#[utoipa::path(
    operation_id = "getPurl",
    tag = "purl",
    params(
        Deprecation,
        ("key" = String, Path, description = "opaque identifier for a fully-qualified PURL, or URL-encoded pURL itself")
    ),
    responses(
        (status = 200, description = "Details for the qualified PURL", body = PurlDetails),
    ),
)]
#[get("/v3/purl/{key}")]
/// Retrieve details of a fully-qualified pURL
pub async fn get(
    service: web::Data<PurlService>,
    db: web::Data<db::ReadOnly>,
    key: web::Path<String>,
    web::Query(Deprecation { deprecated }): web::Query<Deprecation>,
    _: Require<ReadSbom>,
) -> Result<impl Responder, Error> {
    let tx = db.begin().await?;
    if key.starts_with("pkg") {
        let purl = Purl::from_str(&key).map_err(Error::Purl)?;
        Ok(HttpResponse::Ok().json(service.purl_by_purl(&purl, deprecated, &tx).await?))
    } else {
        let id = Uuid::from_str(&key).map_err(|e| Error::IdKey(IdError::InvalidUuid(e)))?;
        Ok(HttpResponse::Ok().json(service.purl_by_uuid(&id, deprecated, &tx).await?))
    }
}

#[utoipa::path(
    operation_id = "listPurl",
    tag = "purl",
    params(
        Query,
        Paginated,
    ),
    responses(
        (status = 200, description = "All relevant matching qualified PURLs", body = PaginatedResults<PurlSummary>),
    ),
)]
#[get("/v3/purl")]
/// List fully-qualified pURLs
pub async fn all(
    service: web::Data<PurlService>,
    db: web::Data<db::ReadOnly>,
    web::Query(search): web::Query<Query>,
    web::Query(paginated): web::Query<Paginated>,
    _: Require<ReadSbom>,
) -> Result<impl Responder, Error> {
    let tx = db.begin().await?;
    Ok(HttpResponse::Ok().json(service.purls(search, paginated, &tx).await?))
}

mod v2 {
    #![allow(deprecated)]
    use super::*;

    #[utoipa::path(
        operation_id = "v2/recommend",
        tag = "purl",
        request_body = RecommendRequest,
        responses(
            (status = 200, description = "Get recommendations and remediations for provided purls", body = RecommendResponse)
        )
    )]
    #[post("/v2/purl/recommend")]
    #[deprecated = "Use the v3 version of this API"]
    pub async fn recommend(
        purl_service: web::Data<PurlService>,
        db: web::Data<db::ReadOnly>,
        request: web::Json<RecommendRequest>,
        _: Require<ReadAdvisory>,
    ) -> Result<impl Responder, Error> {
        let tx = db.begin().await?;
        let recommendations = purl_service.recommend_purls(&request.purls, &tx).await?;

        let response = RecommendResponse { recommendations };

        Ok(HttpResponse::Ok().json(response))
    }
}

mod v3 {
    use super::*;

    #[utoipa::path(
        operation_id = "recommend",
        tag = "purl",
        request_body = RecommendRequest,
        responses(
            (status = 200, description = "Get recommendations and remediations for provided purls", body = RecommendResponse)
        )
    )]
    #[post("/v3/purl/recommend")]
    pub async fn recommend(
        purl_service: web::Data<PurlService>,
        db: web::Data<db::ReadOnly>,
        request: web::Json<RecommendRequest>,
        _: Require<ReadAdvisory>,
    ) -> Result<impl Responder, Error> {
        let tx = db.begin().await?;
        let recommendations = purl_service.recommend_purls(&request.purls, &tx).await?;

        let response = RecommendResponse { recommendations };

        Ok(HttpResponse::Ok().json(response))
    }

    #[utoipa::path(
        operation_id = "recommendReport",
        tag = "purl",
        request_body = RecommendReportRequest,
        responses(
            (status = 200, description = "Aggregated recommendation report for the requested SBOMs", body = RecommendReportResponse),
            (status = 400, description = "Total package count across requested SBOMs exceeds the configured limit"),
        )
    )]
    #[post("/v3/recommend/report")]
    /// Generate an aggregated vendor recommendation report for a set of SBOMs.
    pub async fn recommend_report(
        purl_service: web::Data<PurlService>,
        db: web::Data<db::ReadOnly>,
        request: web::Json<RecommendReportRequest>,
        _: Require<ReadAdvisory>,
    ) -> Result<impl Responder, Error> {
        let tx = db.begin().await?;
        let total = purl_service
            .count_sbom_packages(&request.sbom_ids, &tx)
            .await?;
        if total > purl_service.report_package_limit {
            return Ok(HttpResponse::BadRequest().json(serde_json::json!({
                "error": "package_limit_exceeded",
                "message": format!(
                    "Total packages ({total}) exceeds maximum ({}).",
                    purl_service.report_package_limit
                )
            })));
        }
        let report = purl_service
            .report_for_sboms(&request.sbom_ids, &tx)
            .await?;
        Ok(HttpResponse::Ok().json(report))
    }
}